+++ src/main/kotlin/com/loadingbyte/cinecred/drawer/PageDrawer.kt
@@ -16,7 +16,9 @@ import java.awt.Font
 import java.awt.geom.Path2D
 import java.util.*
 import javax.swing.UIManager
+import kotlin.math.abs
 import kotlin.math.ceil
+import kotlin.math.floor
 
 
 private class StageLayout(val y: Y, val info: DrawnStageInfo)
@@ -44,7 +46,7 @@ fun drawPages(project: Project): List<DrawnPage> {
     val drawnBlocks = drawBlocks(project.styling.contentStyles, textCtx, drawnBodies, blocks)
 
     // Generate a stage image for each stage. These stage images already contain the vertical gaps between the stages.
-    val stageImages = HashMap<Stage, DeferredImage>()
+    var stageImages: MutableMap<Stage, DeferredImage> = HashMap()
     for (page in pages)
         page.stages.withIndex().associateTo(stageImages) { (stageIdx, stage) ->
             val prevStage = page.stages.getOrNull(stageIdx - 1)
@@ -61,26 +63,11 @@ fun drawPages(project: Project): List<DrawnPage> {
         val prelimStageLayouts = HashMap<Stage, StageLayout>()
         for (page in pages)
             prelimStageLayouts.putAll(layoutStages(global, stageImages, page).second)
-
-        // Adjust the stages collected in runtime groups.
-        for (runtimeGroup in runtimeGroups)
-            matchRuntime(
-                pages, stageImages, prelimStageLayouts, pageTopStages, pageBotStages,
-                desiredFrames = runtimeGroup.runtimeFrames,
-                activeStages = runtimeGroup.stages, passiveStages = emptyList()
-            )
-        // If requested, adjust all remaining stages to best achieve the desired overall runtime of the whole sequence.
-        if (global.runtimeFrames.isActive) {
-            val pauseFrames = pages.dropLast(1).sumOf { it.stages.last().style.afterwardSlugFrames }
-            val runtimeStages = runtimeGroups.flatMap(RuntimeGroup::stages)
-            val allStages = stageImages.keys
-            matchRuntime(
-                pages, stageImages, prelimStageLayouts, pageTopStages, pageBotStages,
-                // Just subtract the frames in between pages from the number of desired frames.
-                desiredFrames = global.runtimeFrames.value - pauseFrames,
-                activeStages = allStages - runtimeStages, passiveStages = runtimeStages
-            )
-        }
+        // Use that information to scale the vertical gaps.
+        stageImages = matchRuntime(
+            pages, stageImages, prelimStageLayouts, pageTopStages, pageBotStages,
+            global.runtimeFrames, runtimeGroups
+        )
     }
 
     // Finally, do the real layout pass with potentially changed stage images and combine the stage images
@@ -106,7 +93,7 @@ private fun drawStage(
     // If this stage is a scroll stage is preceded by a card stage, add the vertical gap behind the card stage to the
     // front of this stage because card stage images are not allowed to have extremal vertical gaps baked into them.
     if (stage.style.behavior != CARD && prevStage != null && prevStage.style.behavior == CARD)
-        y += prevStage.vGapAfterPx.toY()
+        y += prevStage.vGapAfterPx.toElasticY()
 
     for ((segmentIdx, segment) in stage.segments.withIndex()) {
         var maxHeight = 0f.toY()
@@ -124,10 +111,7 @@ private fun drawStage(
 
     // If this stage is a scroll stage and not the last stage on the page, add the gap behind it to its image.
     if (stage.style.behavior != CARD && nextStage != null)
-        y += when (nextStage.style.behavior) {
-            CARD -> stage.vGapAfterPx.toY()
-            SCROLL -> stage.vGapAfterPx.toElasticY()
-        }
+        y += stage.vGapAfterPx.toElasticY()
 
     // Set the stage image's height.
     stageImage.height = y
@@ -176,15 +160,13 @@ private fun layoutStages(
     var pageImageHeight = 0f.toY()
     var y = 0f.toY()
     for ((stageIdx, stage) in page.stages.withIndex()) {
-        var stageHeight = stageImages.getValue(stage).height
+        val stageHeight = stageImages.getValue(stage).height
         // Special handling for card stages...
         if (stage.style.behavior == CARD) {
-            // Card stages are rigid and not elastic.
-            val resolvedStageHeight = stageHeight.resolve()
-            stageHeight = resolvedStageHeight.toY()
             // The amount of padding that needs to be added above and below the card's stage image such that
-            // it is centered on the screen.
-            val cardPaddingHeight = (global.heightPx - resolvedStageHeight) / 2f
+            // it is centered on the screen. Notice that as scrolling from/to a card starts/ends at the card's center,
+            // the padding is irrelevant for runtime matching and can hence be rigid.
+            val cardPaddingHeight = (global.heightPx - stageHeight.resolve()) / 2f
             // If this card stage is the first and/or the last stage, make sure that there is extra padding below
             // and above the card stage such that its content is centered vertically.
             if (stageIdx == 0)
@@ -201,9 +183,14 @@ private fun layoutStages(
 
     // Find for each stage:
     //   - If it's a card stage, its middle y coordinate in the overall page image.
-    //   - If it's a scroll stage, the y coordinates in the overall page image that are at the center of the screen
-    //     when the scrolling of this stage starts respectively stops. Also, the number of frames that make up the
-    //     scroll, and a fraction of a frame that should be skipped prior to the beginning of the scroll.
+    //   - If it's a scroll stage:
+    //       - The y coordinates in the overall page image that are at the center of the screen when the scrolling of
+    //         this stage starts respectively stops.
+    //       - Recall that scroll stages can scroll into melted card stages. As such, these coordinates can lie inside
+    //         of card stages. For some purposes however, we need the portion of the scrolled height which is truly
+    //         occupied by the scroll stage.
+    //       - The number of frames that make up the scroll, and a fraction of a frame that should be skipped prior to
+    //         the beginning of the scroll.
     var nextScrollInitAdvance = Float.NaN
     val stageInfo = page.stages.mapIndexed { stageIdx, stage ->
         val (topY, botY) = stageImageBounds[stageIdx]
@@ -222,6 +209,11 @@ private fun layoutStages(
                     SCROLL -> botY - global.heightPx / 2f
                     null -> botY + global.heightPx / 2f
                 }
+                // Find the portion of the scrolled height which really belongs to the scroll stage itself. If you are
+                // confused, recall that scroll stages can scroll into melted card stages.
+                val ownedScrollStartY = if (prevStageBehavior == CARD) topY else scrollStartY
+                val ownedScrollStopY = if (nextStageBehavior == CARD) botY else scrollStopY
+                val ownedScrollHeight = ownedScrollStopY - ownedScrollStartY
                 // Find (a) how much of a single frame to advance the scroll prior to the first frame actually being
                 // rendered (initAdvance) and (b) the number of frames required for the scroll.
                 // To understand this, consider a scenario with the following immediate sequence of scroll stages:
@@ -254,10 +246,10 @@ private fun layoutStages(
                     (scrollStopY.resolve() - scrollStartY.resolve()) / stage.style.scrollPxPerFrame - initAdvance
                 // If fracFrames is just slightly larger than an integer, we want it to be handled as if it was that
                 // integer to compensate for floating point inaccuracy. That is why we have the -0.01 in there.
-                val frames = ceil(fracFrames - 0.01f).toInt()
+                val frames = safeCeil(fracFrames)
                 nextScrollInitAdvance = (-fracFrames).mod(1f)  // Using mod() ensures that the result is positive.
                 // Construct the info object.
-                DrawnStageInfo.Scroll(scrollStartY, scrollStopY, frames, initAdvance)
+                DrawnStageInfo.Scroll(scrollStartY, ownedScrollHeight, frames, initAdvance)
             }
         }
     }
@@ -271,32 +263,264 @@ private fun layoutStages(
 
 private fun matchRuntime(
     pages: List<Page>,
-    stageImages: MutableMap<Stage, DeferredImage>,
-    prelimStageLayouts: Map<Stage, StageLayout>,
+    stageImages: Map<Stage, DeferredImage>,
+    stageLayouts: Map<Stage, StageLayout>,
     pageTopStages: Set<Stage>,
     pageBotStages: Set<Stage>,
-    desiredFrames: Int,
-    activeStages: Collection<Stage>,
-    passiveStages: Collection<Stage>
-) {
-    // Here, we do not store a stretchable length in a Y object, but instead a stretchable number of frames.
-    var frames = 0f.toY()
-
+    globalDesiredFrames: Opt<Int>,
+    runtimeGroups: List<RuntimeGroup>
+): MutableMap<Stage, DeferredImage> {
+    // Prepare two maps which enable simple access to the two stages before and after any scroll stage.
+    val prevStages = HashMap<Stage, Stage>()
+    val nextStages = HashMap<Stage, Stage>()
     for (page in pages)
-        for ((stageIdx, stage) in page.stages.withIndex()) {
-            val prevStage = page.stages.getOrNull(stageIdx - 1)
-            // activate stages -> rigid & elastic  /  passive stages -> rigid only
-            if (stage in activeStages)
-                frames += getStageFrames(prelimStageLayouts, pageTopStages, pageBotStages, stage, prevStage)
-            else if (stage in passiveStages)
-                frames += getStageFrames(prelimStageLayouts, pageTopStages, pageBotStages, stage, prevStage).resolve()
+        for ((scrollIdx, scroll) in page.stages.withIndex())
+            if (scroll.style.behavior == SCROLL) {
+                val prevStage = page.stages.getOrNull(scrollIdx - 1)
+                val nextStage = page.stages.getOrNull(scrollIdx + 1)
+                if (prevStage != null) prevStages[scroll] = prevStage
+                if (nextStage != null) nextStages[scroll] = nextStage
+            }
+
+    // This function removes the card stages from a runtime group and subtract their fixed show/hide runtime from the
+    // runtime group's desired runtime. The function then sorts the group into one of two sets: those whose runtime will
+    // shrink, and those whose runtime will expand. This is required to fulfill a later assumption. Groups whose runtime
+    // is already attained are discarded.
+    val shrinkGroups = HashSet<RuntimeGroup>()
+    val expandGroups = HashSet<RuntimeGroup>()
+    fun addGroup(stages: List<Stage>, frames: Int): RuntimeGroup? {
+        val scrolls = mutableListOf<Stage>()
+        var desiredScrollFrames = frames
+        for (stage in stages)
+            when (stage.style.behavior) {
+                CARD -> desiredScrollFrames -= getCardFrames(pageTopStages, pageBotStages, stage)
+                SCROLL -> scrolls.add(stage)
+            }
+        if (scrolls.isEmpty())
+            return null
+        val group = RuntimeGroup(scrolls.toPersistentList(), desiredScrollFrames)
+        val currentScrollFrames =
+            scrolls.sumOf { scroll -> (stageLayouts.getValue(scroll).info as DrawnStageInfo.Scroll).frames }
+        when {
+            desiredScrollFrames < currentScrollFrames -> shrinkGroups.add(group)
+            desiredScrollFrames > currentScrollFrames -> expandGroups.add(group)
+            // If the desired frames are already attained, discard the group.
+        }
+        return group
+    }
+
+    // Preprocess all user-defined runtime groups accordingly.
+    for (group in runtimeGroups)
+        addGroup(group.stages, group.runtimeFrames)
+
+    // If the global desired runtime is activated, add a runtime group with all remaining stages not covered by
+    // user-defined runtime groups.
+    var globalGroup: RuntimeGroup? = null
+    var globalGroupFrames = 0
+    val remStages = buildList { for (p in pages) for (s in p.stages) if (runtimeGroups.none { s in it.stages }) add(s) }
+    if (globalDesiredFrames.isActive) {
+        // Subtract the desired runtime of the user-defined runtime groups and the runtime in between pages from the
+        // desired global runtime to obtain the desired runtime of the new group.
+        globalGroupFrames = globalDesiredFrames.value -
+                runtimeGroups.sumOf(RuntimeGroup::runtimeFrames) -
+                pages.dropLast(1).sumOf { it.stages.last().style.afterwardSlugFrames }
+        globalGroup = addGroup(remStages, globalGroupFrames)
+    }
+
+    // Run the runtime matching.
+    val (adjustedStageImages, shortfall) =
+        matchRuntime(pages, stageImages, stageLayouts, prevStages, nextStages, shrinkGroups, expandGroups)
+    // If there is no global runtime adjustment happening (either because it is deactivated or because there are no
+    // scroll stages remaining), or if all runtimes were matched perfectly, return the result
+    if (globalGroup == null || shortfall == 0)
+        return adjustedStageImages
+
+    // If we are here, shortfall is non-zero, which means that some stages did not exactly attain their desired runtime.
+    // Apart from some very rare situations where an exact match is impossible, this means that some runtime groups have
+    // run out of vertical gaps to expense for decreasing their runtime.
+    // However, the global runtime group expects all other runtime groups to have exact matches. Thus, we need to
+    // compensate for the other groups' shortfall in the global group's desired runtime.
+    shrinkGroups.remove(globalGroup)
+    expandGroups.remove(globalGroup)
+    addGroup(remStages, globalGroupFrames + shortfall)
+    return matchRuntime(pages, stageImages, stageLayouts, prevStages, nextStages, shrinkGroups, expandGroups).first
+}
+
+
+private fun matchRuntime(
+    pages: List<Page>,
+    stageImages: Map<Stage, DeferredImage>,
+    stageLayouts: Map<Stage, StageLayout>,
+    prevStages: Map<Stage, Stage>,
+    nextStages: Map<Stage, Stage>,
+    shrinkGroups: Set<RuntimeGroup>,
+    expandGroups: Set<RuntimeGroup>
+): Pair<MutableMap<Stage, DeferredImage>, Int> {
+    // Naturally, we do not modify the vertical gaps of standalone cards. Further, we only adjust the gaps of a melted
+    // card if it either borders exactly one scroll stage (in which case it is basically treated exactly the same as the
+    // scroll stage) or if it borders two scroll stages which both need to shrink or both need to expand (in which case
+    // the card will shrink/expand as far as the more modest of the two scroll stages).
+    // To enforce this policy, we "freeze" the vertical height of all stages that will not be adjusted, that is, we make
+    // the stages' heights rigid and remember that we've frozen the stages.
+    val frozenCards = HashSet<Stage>()
+    for (page in pages)
+        for ((cardIdx, card) in page.stages.withIndex()) {
+            if (card.style.behavior == CARD) {
+                operator fun Iterable<RuntimeGroup>.contains(stage: Stage) = any { stage in it.stages }
+                val prevStage = page.stages.getOrNull(cardIdx - 1)
+                val nextStage = page.stages.getOrNull(cardIdx + 1)
+                if (// Case 1: The card is standalone and not melted with any scroll stage.
+                    prevStage == null && nextStage == null ||
+                    // Case 2: The card is melted on both sides, and the scrolls on both sides must neither both shrink
+                    //         nor both expand.
+                    prevStage != null && nextStage != null &&
+                    (prevStage !in shrinkGroups || nextStage !in shrinkGroups) &&
+                    (prevStage !in expandGroups || nextStage !in expandGroups)
+                ) {
+                    frozenCards.add(card)
+                    stageImages.getValue(card).apply { height = height.resolve().toY() }
+                }
+            }
         }
 
-    val elasticScaling = frames.deresolve(desiredFrames.toFloat())
+    // Now comes the real deal: for the groups which shrink and for those who expand separately, we collectively match
+    // the desired runtime.
+    val img = HashMap(stageImages)
+    var shortfall = 0
+    shortfall += matchRuntimeInOneDir(stageLayouts, prevStages, nextStages, img, HashSet(shrinkGroups), frozenCards)
+    shortfall += matchRuntimeInOneDir(stageLayouts, prevStages, nextStages, img, HashSet(expandGroups), frozenCards)
+    return Pair(img, shortfall)
+}
 
-    for (stage in activeStages)
-        if (stage.style.behavior == SCROLL)
-            stageImages[stage] = stageImages.getValue(stage).copy(elasticScaling = elasticScaling)
+// Note: "direction" refers to either shrinking or expanding.
+private fun matchRuntimeInOneDir(
+    stageLayouts: Map<Stage, StageLayout>,
+    prevStages: Map<Stage, Stage>,
+    nextStages: Map<Stage, Stage>,
+    // Manipulated collections:
+    stageImages: MutableMap<Stage, DeferredImage>,
+    groups: MutableSet<RuntimeGroup>,
+    frozenCards: MutableSet<Stage>
+): Int {
+    fun prevCard(scroll: Stage) = prevStages[scroll]?.let { prev -> if (prev.style.behavior == CARD) prev else null }
+    fun nextCard(scroll: Stage) = nextStages[scroll]?.let { next -> if (next.style.behavior == CARD) next else null }
+
+    // If the given card has not been frozen yet, this function first scales its vertical gaps and then freezes it
+    // similar to before, i.e., makes the height of the card's image rigid.
+    fun tryFreezeCard(card: Stage, elasticScaling: Float) {
+        if (card !in frozenCards) {
+            frozenCards.add(card)
+            stageImages[card] = stageImages.getValue(card).copy(elasticScaling = elasticScaling)
+                .apply { height = height.resolve().toY() }
+        }
+    }
+
+    var shortfall = 0
+    // Until all runtime groups have been matched...
+    while (groups.isNotEmpty()) {
+        // Find the group which needs the most modest elastic scaling to match its desired runtime. Then apply that
+        // scaling to the group and all unfrozen adjacent cards (which are consequently frozen). By iteratively
+        // advancing to the more and more extreme elastic scaling, we fulfill the policy that each card should be
+        // scaled in the same way as the more modest of its two adjacent scroll stages.
+        var modestElasticScaling = Float.NaN
+        var modestGroup: RuntimeGroup? = null
+        var modestDiscreteFrames = -1
+        for (group in groups) {
+            // Here, we do not store a stretchable length in a Y object, but instead a stretchable number of frames.
+            // For each continuous and uninterrupted stretch of adjacent scroll stages, find the number of frames needed
+            // by that stretch. Be aware that we assume that group.stages contains the stages in the same order as they
+            // appear in the credits.
+            val stretchFrames = mutableListOf<Y>()
+            for (scroll in group.stages) {
+                val prevStage = prevStages[scroll]
+                // If the previous stretch ended, start a new one.
+                if (prevStage == null || prevStage.style.behavior != SCROLL || prevStage !in group.stages)
+                    stretchFrames.add(0f.toY())
+                // The number of frames needed for this scroll stage is the vertical scrolled distance divided by the
+                // scrolling speed. Apart from space owned by the scroll stage itself, the distance includes half the
+                // space of adjacent cards. Although the precomputed scrollStartY and scrollStopY values could already
+                // provide access to this distance, we need to account for the dynamically frozen card heights. As such,
+                // we manually collect the scrolled distance here.
+                var scrolledDist = (stageLayouts.getValue(scroll).info as DrawnStageInfo.Scroll).ownedScrollHeight
+                prevCard(scroll)?.let { prevCard -> scrolledDist += stageImages.getValue(prevCard).height / 2f }
+                nextCard(scroll)?.let { nextCard -> scrolledDist += stageImages.getValue(nextCard).height / 2f }
+                var frames = stretchFrames.last() + scrolledDist / scroll.style.scrollPxPerFrame
+                // The scroll frames we just computed do not contain the frame at DrawnStageInfo.Scroll.scrollStopY.
+                // For example, if there is a scroll height "scrollStopY - scrollStartY" of 10 pixels, and we scroll
+                // 2 pixels per frame, the result is 5 frames (at the pixel offsets 0, 2, 4, 6, 8): the first frame at
+                // the top is included, but the last frame at the bottom is not.
+                // However, if the stage is not preceded by another scroll stage, but instead either by void or by a
+                // card stage, we need to also drop the frame at scrollStartY. As a result, the scroll now contains only
+                // "moving" frames along the continuous stretch of scroll stages, but not the still frames at the
+                // stretch's start and end.
+                if (prevStage == null || prevStage.style.behavior != SCROLL)
+                    frames -= 1f
+                stretchFrames[stretchFrames.lastIndex] = frames
+            }
+
+            // Find the elastic scaling which, when applied, makes the runtime group attain its desired runtime.
+            var elasticScaling = stretchFrames.reduce(Y::plus).deresolve(group.runtimeFrames.toFloat())
+
+            // The above operation is only correct if there is only one continuous stretch of scroll stages in the
+            // runtime group. If there are however two or more stretches, the total number of frames can no longer be
+            // found by just adding up the fractional frames of all stretches. Instead, before the addition, there has
+            // to be a ceiling operation applied to each stretch individually to discretize the stretch's frames. This
+            // ceiling operation would turn the Y curve that models a stretch's frames into a stepwise function. All
+            // stepwise functions would then be added together to an overall stepwise function. As this is difficult to
+            // implement explicitly, we instead go another route.
+            // We first estimate the elastic scaling using the regular deresolve() function above. Then, we find the
+            // actual number of discrete frames under that elastic scaling using the ceiling procedure outlined above.
+            // If it doesn't match the desired frames, we take steps up or down the outline overall stepwise function
+            // until the desired frames are matched.
+            // Notice that due to the stepwise nature of the overall function, there are some very special cases in
+            // which the desired frames can't be attained exactly; in those cases, we just stop when we have surpassed
+            // the desired frames.
+            var discreteFrames = stretchFrames.sumOf { frames -> safeCeil(frames.resolve(elasticScaling)) }
+            if (stretchFrames.size > 1) {
+                // Find whether we need to increase or decrease the elastic scaling.
+                val up = discreteFrames < group.runtimeFrames
+                var ctr = 0  // failsafe
+                while (
+                    ctr++ < 100 &&
+                    // Stop if we try to further decrease an already 0 elastic scaling.
+                    (up || elasticScaling != 0f) &&
+                    // Stop as soon as the desired runtime is attained (or in some rare cases overshot).
+                    if (up) discreteFrames < group.runtimeFrames else discreteFrames > group.runtimeFrames
+                ) {
+                    // Take one step up or down the overall stepwise function. We do this by adjusting the elastic
+                    // scaling such that just one of the stretch's stepwise functions (which make up the overall
+                    // function) takes a single step.
+                    elasticScaling = when (up) {
+                        true -> stretchFrames.minOf { it.deresolve(floor(it.resolve(elasticScaling) + 1f)) }
+                        false -> stretchFrames.maxOf { it.deresolve(ceil(it.resolve(elasticScaling) - 1f)) }
+                    }
+                    discreteFrames = stretchFrames.sumOf { frames -> safeCeil(frames.resolve(elasticScaling)) }
+                }
+            }
+
+            if (modestElasticScaling.isNaN() || abs(elasticScaling - 1f) < abs(modestElasticScaling - 1f)) {
+                modestElasticScaling = elasticScaling
+                modestGroup = group
+                modestDiscreteFrames = discreteFrames
+            }
+        }
+
+        // As stated before, we now apply the most modest elastic scaling to the corresponding group and its adjacent
+        // unfrozen card stages, freeze those cards, and take the group out of the equation as it's now matched.
+        groups.remove(modestGroup)
+        for (scroll in modestGroup!!.stages) {
+            stageImages[scroll] = stageImages.getValue(scroll).copy(elasticScaling = modestElasticScaling)
+            prevCard(scroll)?.let { prevCard -> tryFreezeCard(prevCard, modestElasticScaling) }
+            nextCard(scroll)?.let { nextCard -> tryFreezeCard(nextCard, modestElasticScaling) }
+        }
+
+        // Record by how many frames the actually attained runtime diverges from the group's desired one. If there is a
+        // divergence, it either stems from the very rare impossibilities outline above, or more likely from a group
+        // which does not have enough vertical gaps to expend in order to fully attain a very low desired runtime.
+        shortfall += modestGroup.runtimeFrames - modestDiscreteFrames
+    }
+
+    return shortfall
 }
 
 
@@ -381,19 +605,6 @@ private fun DeferredImage.drawMeltedCardArrowGuide(global: Global, y: Y) {
 }
 
 
-private fun getStageFrames(
-    stageLayouts: Map<Stage, StageLayout>,
-    pageTopStages: Set<Stage>,
-    pageBotStages: Set<Stage>,
-    stage: Stage,
-    prevStage: Stage?
-): Y =
-    when (stage.style.behavior) {
-        CARD -> getCardFrames(pageTopStages, pageBotStages, stage).toFloat().toY()
-        SCROLL -> getScrollFrames(stageLayouts, stage, prevStage)
-    }
-
-
 private fun getCardFrames(
     pageTopStages: Set<Stage>,
     pageBotStages: Set<Stage>,
@@ -409,20 +620,8 @@ private fun getCardFrames(
 }
 
 
-private fun getScrollFrames(
-    stageLayouts: Map<Stage, StageLayout>,
-    stage: Stage,
-    prevStage: Stage?
-): Y {
-    val stageInfo = stageLayouts.getValue(stage).info as DrawnStageInfo.Scroll
-    // Notice that the scroll frames do not contain the frame at scrollStopY. For example, if there is a scroll height
-    // "scrollStopY - scrollStartY" of 10 pixels, and we scroll 2 pixels per frame, the first frame at the top is
-    // included, but the last frame at the bottom is not.
-    var frames = (stageInfo.scrollStopY - stageInfo.scrollStartY) / stage.style.scrollPxPerFrame
-    // If the stage is not preceded by another scroll stage, but either by void or by a card stage, also drop the frame
-    // at scrollStartY. As a result, the scroll now contains only "moving" frames along the continuous stretch of scroll
-    // stages, but not the still frames at its start and end.
-    if (prevStage == null || prevStage.style.behavior != SCROLL)
-        frames -= 1f
-    return frames
-}
+/**
+ * If [x] is just slightly larger than an integer, we want it to be handled as if it was that integer to compensate for
+ * floating point inaccuracy. That is why we have the slight offset in there.
+ */
+private fun safeCeil(x: Float): Int = ceil(x - 0.01f).toInt()